A static version of the interface
so that margot can work without having to worry about django or editing
the javascript in the mess of simsam guts. instead, this shows what
simsam should be rendered to look like by django so she can just play
with the css and interface-related javascript in a way that can be
easily re-imported.