cleanup a first object bug
I think this was a result of an ordering effect in corruptor, where the
object was added then Sam was killed? Any new project would correctly
crop, but would not write the freshly cropped object to the animation.
Then Sam would re-appear instead of cropped object #0. Not sure, but
these changes appear to make that problem go away.